Update README.md
1. Title : hotel rooms booking Management System
2. Abstract : This C programming project involves the creation of hotel rooms bookings 
database management system. The system allows the user to findout the availabity of rooms 
with different accomodations. It includes information such as accommodation types , num of 
guests, no.of hrs spent . The system stores the records in a file, which can be easily accessed 
and modified. The project is implemented using the concepts of file handling data structures 
in C programming language. The user interface is designed to be userfriendly. The project is 
useful for managing in hotels and to know total payment bill. 
3. Introduction : The C programming language has been used for a wide range of 
applications, including the development of databases.. The system allows the user to findout 
the availability of rooms with different accomodations. This database management system 
will be designed to be user-friendly, reliable, and efficient, using various C programming 
concepts such as data structures, file handling, and memory allocation. User can sort num of 
guests , types of accomodations and num of hours.This system can help to streamline 
administrative tasks such as record-keeping and data analysis. Furthermore, the project will 
provide an opportunity to apply and reinforce key programming concepts and techniques in C
language. 
3.1. Traditional Method : From a very long time we have been using the traditional method 
of pen and paper system for hotel room bookings by writing every single detail of hotel room 
in records. 
3.2. Problems using Traditional method : Using a pen and paper method for hotel rooms 
database can lead to several problems, including:
1. Time-consuming: Maintaining hotel rooms booking data base using a pen and paper 
method can be time-consuming, as it involves manually writing, organizing, and storing the 
information. This process can be especially challenging if there are many bookings to keep 
track of or if the database needs to be updated frequently.
 2. Prone to errors: Using a pen and paper method can lead to errors, such as misspellings, 
inaccurate data, or misplaced records. These errors can be time-consuming and costly to 
correct and may lead to incorrect decisions being made based on the faulty data.
3. Limited accessibility: A pen and paper method of hotel room bookings database is not 
easily accessible to multiple users at once. Only one person can access the records at a time, 
leading to delays and inefficiencies in decision-making.
 4. Lack of security: Using a pen and paper method the records can be easily misplaced or 
stolen.
 5. Difficult to analyze: Analyzing data using a pen and paper method can be difficult, as it 
requires manually sifting through records and calculating total bill. This can be 
timeconsuming and prone to errors, making it challenging to gain valuable insights from the 
data. Overall, using a pen and paper method for artist database can be inefficient, prone to 
errors, and time-consuming. A digital database management system can help to address these 
issues and provide a more efficient, secure, and accessible way of managing room bookings.
3.3. Solution : So here is a C program to manage a hotel room bookings database which can
provide an efficient, secure, and accessible way of managing total bill , eliminating the need 
for the pen and paper method. The program can be customized to meet the specific